Move wrapping on semicolon from `wrapping` rule to `statement-wrapping` rule
