+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/rules/WrappingRule.kt
@@ -11,7 +11,6 @@ import com.pinterest.ktlint.rule.engine.core.api.ElementType.COMMA
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.CONDITION
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.DESTRUCTURING_DECLARATION
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.DOT
-import com.pinterest.ktlint.rule.engine.core.api.ElementType.ENUM_ENTRY
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.EOL_COMMENT
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.FUN
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.FUNCTION_LITERAL
@@ -27,7 +26,6 @@ import com.pinterest.ktlint.rule.engine.core.api.ElementType.OBJECT_LITERAL
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.RBRACE
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.RBRACKET
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.RPAR
-import com.pinterest.ktlint.rule.engine.core.api.ElementType.SEMICOLON
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.STRING_TEMPLATE
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.SUPER_TYPE_CALL_ENTRY
 import com.pinterest.ktlint.rule.engine.core.api.ElementType.SUPER_TYPE_ENTRY
@@ -69,7 +67,6 @@ import com.pinterest.ktlint.rule.engine.core.api.nextCodeLeaf
 import com.pinterest.ktlint.rule.engine.core.api.nextCodeSibling
 import com.pinterest.ktlint.rule.engine.core.api.nextLeaf
 import com.pinterest.ktlint.rule.engine.core.api.nextSibling
-import com.pinterest.ktlint.rule.engine.core.api.noNewLineInClosedRange
 import com.pinterest.ktlint.rule.engine.core.api.prevCodeLeaf
 import com.pinterest.ktlint.rule.engine.core.api.prevLeaf
 import com.pinterest.ktlint.rule.engine.core.api.prevSibling
@@ -143,7 +140,6 @@ public class WrappingRule :
             ARROW -> rearrangeArrow(node, autoCorrect, emit)
             WHITE_SPACE -> line += node.text.count { it == '\n' }
             CLOSING_QUOTE -> rearrangeClosingQuote(node, autoCorrect, emit)
-            SEMICOLON -> insertNewLineAfterSemi(node, autoCorrect, emit)
         }
     }
 
@@ -539,34 +535,6 @@ public class WrappingRule :
         val r = node.nextSibling { it.elementType == RBRACE } ?: return
         if (!r.prevLeaf().isWhiteSpaceWithNewline()) {
             requireNewlineBeforeLeaf(r, autoCorrect, emit, node.indent())
-        }
-    }
-
-    private fun insertNewLineAfterSemi(
-        node: ASTNode,
-        autoCorrect: Boolean,
-        emit: (offset: Int, errorMessage: String, canBeAutoCorrected: Boolean) -> Unit,
-    ) {
-        val previousCodeLeaf = node.prevCodeLeaf()?.lastChildLeafOrSelf() ?: return
-        val nextCodeLeaf = node.nextCodeLeaf()?.firstChildLeafOrSelf() ?: return
-        if (previousCodeLeaf.treeParent.elementType == ENUM_ENTRY && nextCodeLeaf.elementType == RBRACE) {
-            // Allow
-            // enum class INDEX2 { ONE, TWO, THREE; }
-            return
-        }
-        if (noNewLineInClosedRange(previousCodeLeaf, nextCodeLeaf)) {
-            requireNewlineAfterLeaf(node, autoCorrect, emit, previousCodeLeaf.indent())
-            node
-                .treeParent
-                .takeIf { it.elementType == BLOCK }
-                ?.let { block ->
-                    beforeVisitBlock(block, autoCorrect, emit)
-                    block
-                        .treeParent
-                        .takeIf { it.elementType == FUNCTION_LITERAL }
-                        ?.findChildByType(ARROW)
-                        ?.let { arrow -> rearrangeArrow(arrow, autoCorrect, emit) }
-                }
         }
     }
