+++ ktlint-ruleset-standard/src/test/kotlin/com/pinterest/ktlint/ruleset/standard/rules/StatementWrappingRuleTest.kt
@@ -2,7 +2,7 @@ package com.pinterest.ktlint.ruleset.standard.rules
 
 import com.pinterest.ktlint.test.KtLintAssertThat
 import com.pinterest.ktlint.test.LintViolation
-import org.junit.jupiter.api.Disabled
+import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 
 class StatementWrappingRuleTest {
@@ -23,7 +23,7 @@ class StatementWrappingRuleTest {
             }
             """.trimIndent()
         statementWrappingRuleAssertThat(code)
-            .hasLintViolation(1, 13, "Expected new line after '{'")
+            .hasLintViolation(1, 13, "Missing newline after '{'")
             .isFormattedAs(formattedCode)
     }
 
@@ -41,7 +41,7 @@ class StatementWrappingRuleTest {
             }
             """.trimIndent()
         statementWrappingRuleAssertThat(code)
-            .hasLintViolation(2, 19, "Expected new line before '}'")
+            .hasLintViolation(2, 19, "Missing newline before '}'")
             .isFormattedAs(formattedCode)
     }
 
@@ -70,8 +70,8 @@ class StatementWrappingRuleTest {
             """.trimIndent()
         statementWrappingRuleAssertThat(code)
             .hasLintViolations(
-                LintViolation(1, 13, "Expected new line after '{'"),
-                LintViolation(1, 27, "Expected new line before '}'"),
+                LintViolation(1, 13, "Missing newline after '{'"),
+                LintViolation(1, 27, "Missing newline before '}'"),
             ).isFormattedAs(formattedCode)
     }
 
@@ -99,10 +99,10 @@ class StatementWrappingRuleTest {
             """.trimIndent()
         statementWrappingRuleAssertThat(code)
             .hasLintViolations(
-                LintViolation(1, 13, "Expected new line after '{'"),
-                LintViolation(2, 19, "Expected new line before '}'"),
-                LintViolation(3, 13, "Expected new line after '{'"),
-                LintViolation(5, 6, "Expected new line before '}'"),
+                LintViolation(1, 13, "Missing newline after '{'"),
+                LintViolation(2, 19, "Missing newline before '}'"),
+                LintViolation(3, 13, "Missing newline after '{'"),
+                LintViolation(5, 6, "Missing newline before '}'"),
             ).isFormattedAs(formattedCode)
     }
 
@@ -115,30 +115,6 @@ class StatementWrappingRuleTest {
             fun foo3() { /* no-op */ }
             """.trimIndent()
         statementWrappingRuleAssertThat(code).hasNoLintViolations()
-    }
-
-    @Disabled
-    @Test
-    fun `Given a block containing multiple statements on same line separated by a semi colon`() {
-        val code =
-            """
-            fun foo() {
-                doSomething(); doSomething()
-            }
-            """.trimIndent()
-        val formattedCode =
-            """
-            fun foo() {
-                doSomething()
-                doSomething()
-            }
-            """.trimIndent()
-        statementWrappingRuleAssertThat(code)
-            .hasLintViolations(
-                LintViolation(1, 13, "Expected new line after '{'"),
-                LintViolation(1, 28, "Expected new line after ';'"),
-                LintViolation(1, 40, "Expected new line before '}'"),
-            ).isFormattedAs(formattedCode)
     }
 
     @Test
@@ -169,7 +145,7 @@ class StatementWrappingRuleTest {
             }
             """.trimIndent()
         statementWrappingRuleAssertThat(code)
-            .hasLintViolation(1, 12, "Expected new line after '{'")
+            .hasLintViolation(1, 12, "Missing newline after '{'")
             .isFormattedAs(formattedCode)
     }
 
@@ -187,7 +163,7 @@ class StatementWrappingRuleTest {
             }
             """.trimIndent()
         statementWrappingRuleAssertThat(code)
-            .hasLintViolation(2, 18, "Expected new line before '}'")
+            .hasLintViolation(2, 18, "Missing newline before '}'")
             .isFormattedAs(formattedCode)
     }
 
@@ -211,8 +187,8 @@ class StatementWrappingRuleTest {
             """.trimIndent()
         statementWrappingRuleAssertThat(code)
             .hasLintViolations(
-                LintViolation(2, 16, "Expected new line after '{'"),
-                LintViolation(3, 20, "Expected new line before '}'"),
+                LintViolation(2, 16, "Missing newline after '{'"),
+                LintViolation(3, 20, "Missing newline before '}'"),
             ).isFormattedAs(formattedCode)
     }
 
@@ -236,8 +212,8 @@ class StatementWrappingRuleTest {
             """.trimIndent()
         statementWrappingRuleAssertThat(code)
             .hasLintViolations(
-                LintViolation(2, 24, "Expected new line after '{'"),
-                LintViolation(3, 23, "Expected new line before '}'"),
+                LintViolation(2, 24, "Missing newline after '{'"),
+                LintViolation(3, 23, "Missing newline before '}'"),
             ).isFormattedAs(formattedCode)
     }
 
@@ -261,8 +237,8 @@ class StatementWrappingRuleTest {
             """.trimIndent()
         statementWrappingRuleAssertThat(code)
             .hasLintViolations(
-                LintViolation(2, 26, "Expected new line after '{'"),
-                LintViolation(3, 23, "Expected new line before '}'"),
+                LintViolation(2, 26, "Missing newline after '{'"),
+                LintViolation(3, 23, "Missing newline before '}'"),
             ).isFormattedAs(formattedCode)
     }
 
@@ -286,8 +262,8 @@ class StatementWrappingRuleTest {
             """.trimIndent()
         statementWrappingRuleAssertThat(code)
             .hasLintViolations(
-                LintViolation(2, 10, "Expected new line after '{'"),
-                LintViolation(3, 23, "Expected new line before '}'"),
+                LintViolation(2, 10, "Missing newline after '{'"),
+                LintViolation(3, 23, "Missing newline before '}'"),
             ).isFormattedAs(formattedCode)
     }
 
@@ -316,9 +292,9 @@ class StatementWrappingRuleTest {
             """.trimIndent()
         statementWrappingRuleAssertThat(code)
             .hasLintViolations(
-                LintViolation(2, 11, "Expected new line after '{'"),
-                LintViolation(3, 30, "Expected new line after '{'"),
-                LintViolation(4, 17, "Expected new line after '{'"),
+                LintViolation(2, 11, "Missing newline after '{'"),
+                LintViolation(3, 30, "Missing newline after '{'"),
+                LintViolation(4, 17, "Missing newline after '{'"),
             ).isFormattedAs(formattedCode)
     }
 
@@ -347,9 +323,9 @@ class StatementWrappingRuleTest {
             """.trimIndent()
         statementWrappingRuleAssertThat(code)
             .hasLintViolations(
-                LintViolation(3, 23, "Expected new line before '}'"),
-                LintViolation(4, 23, "Expected new line before '}'"),
-                LintViolation(5, 23, "Expected new line before '}'"),
+                LintViolation(3, 23, "Missing newline before '}'"),
+                LintViolation(4, 23, "Missing newline before '}'"),
+                LintViolation(5, 23, "Missing newline before '}'"),
             ).isFormattedAs(formattedCode)
     }
 
@@ -384,9 +360,9 @@ class StatementWrappingRuleTest {
             """.trimIndent()
         statementWrappingRuleAssertThat(code)
             .hasLintViolations(
-                LintViolation(4, 23, "Expected new line before '}'"),
-                LintViolation(6, 22, "Expected new line before '}'"),
-                LintViolation(8, 21, "Expected new line before '}'"),
+                LintViolation(4, 23, "Missing newline before '}'"),
+                LintViolation(6, 22, "Missing newline before '}'"),
+                LintViolation(8, 21, "Missing newline before '}'"),
             ).isFormattedAs(formattedCode)
     }
 
@@ -422,8 +398,8 @@ class StatementWrappingRuleTest {
             """.trimIndent()
         statementWrappingRuleAssertThat(code)
             .hasLintViolations(
-                LintViolation(1, 14, "Expected new line after '{'"),
-                LintViolation(6, 24, "Expected new line after '{'"),
+                LintViolation(1, 14, "Missing newline after '{'"),
+                LintViolation(6, 24, "Missing newline after '{'"),
             ).isFormattedAs(formattedCode)
     }
 
@@ -459,8 +435,8 @@ class StatementWrappingRuleTest {
             """.trimIndent()
         statementWrappingRuleAssertThat(code)
             .hasLintViolations(
-                LintViolation(4, 7, "Expected new line before '}'"),
-                LintViolation(8, 27, "Expected new line before '}'"),
+                LintViolation(4, 7, "Missing newline before '}'"),
+                LintViolation(8, 27, "Missing newline before '}'"),
             ).isFormattedAs(formattedCode)
     }
 
@@ -492,7 +468,7 @@ class StatementWrappingRuleTest {
                 }
             """.trimIndent()
         statementWrappingRuleAssertThat(code)
-            .hasLintViolation(2, 18, "Expected new line after '->'")
+            .hasLintViolation(2, 18, "Missing newline after '->'")
             .isFormattedAs(formattedCode)
     }
 
@@ -503,5 +479,748 @@ class StatementWrappingRuleTest {
             enum class FooBar { FOO, BAR }
             """.trimIndent()
         statementWrappingRuleAssertThat(code).hasNoLintViolations()
+    }
+
+    @Nested
+    inner class `Issue 1078 - Given multiple expression seperated with semi in a single line` {
+        @Nested
+        inner class `Given multiple variables` {
+            @Test
+            fun `Given two variables`() {
+                val code =
+                    """
+                    fun foo() {
+                        val bar1 = 3; val bar2 = 2
+                        val fooBar1: String = ""; val fooBar2: () -> Unit = {  }
+                    }
+                    """.trimIndent()
+                val formattedCode =
+                    """
+                    fun foo() {
+                        val bar1 = 3
+                        val bar2 = 2
+                        val fooBar1: String = ""
+                        val fooBar2: () -> Unit = {  }
+                    }
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .addAdditionalRuleProvider { NoSemicolonsRule() }
+                    .hasLintViolations(
+                        LintViolation(2, 18, "Missing newline after ';'"),
+                        LintViolation(3, 30, "Missing newline after ';'"),
+                    ).isFormattedAs(formattedCode)
+            }
+
+            @Test
+            fun `Given two variables run without NoSemicolonsRule`() {
+                val code =
+                    """
+                    fun foo() {
+                        val bar1 = 3; val bar2 = 2
+                        val fooBar1: String = ""; val fooBar2: () -> Unit = {  }
+                    }
+                    """.trimIndent()
+                val formattedCode =
+                    """
+                    fun foo() {
+                        val bar1 = 3;
+                        val bar2 = 2
+                        val fooBar1: String = "";
+                        val fooBar2: () -> Unit = {  }
+                    }
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .hasLintViolations(
+                        LintViolation(2, 18, "Missing newline after ';'"),
+                        LintViolation(3, 30, "Missing newline after ';'"),
+                    ).isFormattedAs(formattedCode)
+            }
+
+            @Test
+            fun `Given more than two variables`() {
+                val code =
+                    """
+                    fun foo() {
+                        val bar1 = 3; val bar2 = 2; val bar3 = 3; val bar4: () -> Unit = {  }; val bar4: String = "";
+                    }
+                    """.trimIndent()
+                val formattedCode =
+                    """
+                    fun foo() {
+                        val bar1 = 3
+                        val bar2 = 2
+                        val bar3 = 3
+                        val bar4: () -> Unit = {  }
+                        val bar4: String = ""
+                    }
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .addAdditionalRuleProvider { NoSemicolonsRule() }
+                    .hasLintViolations(
+                        LintViolation(2, 18, "Missing newline after ';'"),
+                        LintViolation(2, 32, "Missing newline after ';'"),
+                        LintViolation(2, 46, "Missing newline after ';'"),
+                        LintViolation(2, 75, "Missing newline after ';'"),
+                    ).isFormattedAs(formattedCode)
+            }
+
+            @Test
+            fun `Given variables with comments`() {
+                val code =
+                    """
+                    fun foo() {
+                        val bar1 = 3; val bar2 = 2; // this is end comment
+                        val bar1 = 3; /* block comment */ val bar2 = 2;
+                    }
+                    """.trimIndent()
+                val formattedCode =
+                    """
+                    fun foo() {
+                        val bar1 = 3
+                        val bar2 = 2; // this is end comment
+                        val bar1 = 3
+                        /* block comment */ val bar2 = 2
+                    }
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .addAdditionalRuleProvider { NoSemicolonsRule() }
+                    .hasLintViolations(
+                        LintViolation(2, 18, "Missing newline after ';'"),
+                        LintViolation(3, 18, "Missing newline after ';'"),
+                    ).isFormattedAs(formattedCode)
+            }
+        }
+
+        @Nested
+        inner class `Given multiple classes, functions and init blocks` {
+            @Test
+            fun `Given multiple function declaration`() {
+                val code =
+                    """
+                    public fun foo1() {
+                        // no-op
+                    }; public fun foo2() {
+                        // no-op
+                    }; fun foo3() = 0
+
+                    public fun foo4() = 1; public fun foo5() {
+                        // no-op
+                    }
+                    """.trimIndent()
+                val formattedCode =
+                    """
+                    public fun foo1() {
+                        // no-op
+                    }
+                    public fun foo2() {
+                        // no-op
+                    }
+                    fun foo3() = 0
+
+                    public fun foo4() = 1
+                    public fun foo5() {
+                        // no-op
+                    }
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .addAdditionalRuleProvider { NoSemicolonsRule() }
+                    .hasLintViolations(
+                        LintViolation(3, 3, "Missing newline after ';'"),
+                        LintViolation(5, 3, "Missing newline after ';'"),
+                        LintViolation(7, 23, "Missing newline after ';'"),
+                    ).isFormattedAs(formattedCode)
+            }
+
+            @Test
+            fun `Given multiple function declaration with comments`() {
+                val code =
+                    """
+                    public fun foo1() {
+                        // no-op
+                    }; /* block comment */ public fun foo2() {
+                        // no-op
+                    }; fun foo3() = 0 // single line comment
+                    """.trimIndent()
+                val formattedCode =
+                    """
+                    public fun foo1() {
+                        // no-op
+                    }
+                    /* block comment */ public fun foo2() {
+                        // no-op
+                    }
+                    fun foo3() = 0 // single line comment
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .addAdditionalRuleProvider { NoSemicolonsRule() }
+                    .hasLintViolations(
+                        LintViolation(3, 3, "Missing newline after ';'"),
+                        LintViolation(5, 3, "Missing newline after ';'"),
+                    ).isFormattedAs(formattedCode)
+            }
+
+            @Test
+            fun `Given multiple function invocations`() {
+                val code =
+                    """
+                    class Bar {
+                        public fun foo1() = 0
+                        fun foo2() = 0
+                        fun foo3(lambda: () -> Unit) = 0
+
+                        init {
+                            foo1(); foo3 {  }; foo2()
+                        }
+                    }
+                    """.trimIndent()
+                val formattedCode =
+                    """
+                    class Bar {
+                        public fun foo1() = 0
+                        fun foo2() = 0
+                        fun foo3(lambda: () -> Unit) = 0
+
+                        init {
+                            foo1()
+                            foo3 {  }
+                            foo2()
+                        }
+                    }
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .addAdditionalRuleProvider { NoSemicolonsRule() }
+                    .hasLintViolations(
+                        LintViolation(7, 16, "Missing newline after ';'"),
+                        LintViolation(7, 27, "Missing newline after ';'"),
+                    ).isFormattedAs(formattedCode)
+            }
+
+            @Test
+            fun `Given a multiline class declaration`() {
+                val code =
+                    """
+                    public class FooBar1 {
+
+                    }; public class FooBar2 {
+
+                    }
+
+                    public class FooBar3; public class FooBar4
+                    """.trimIndent()
+                val formattedCode =
+                    """
+                    public class FooBar1 {
+
+                    }
+                    public class FooBar2 {
+
+                    }
+
+                    public class FooBar3
+                    public class FooBar4
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .addAdditionalRuleProvider { NoSemicolonsRule() }
+                    .hasLintViolations(
+                        LintViolation(3, 3, "Missing newline after ';'"),
+                        LintViolation(7, 22, "Missing newline after ';'"),
+                    ).isFormattedAs(formattedCode)
+            }
+
+            @Test
+            fun `Given a multiline class declaration with comments`() {
+                val code =
+                    """
+                    public class FooBar1 {
+
+                    }; /* block comment */ public class FooBar2 {
+
+                    }; public class FooBar2 {
+
+                    } // single line comment
+                    """.trimIndent()
+                val formattedCode =
+                    """
+                    public class FooBar1 {
+
+                    }
+                    /* block comment */ public class FooBar2 {
+
+                    }
+                    public class FooBar2 {
+
+                    } // single line comment
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .addAdditionalRuleProvider { NoSemicolonsRule() }
+                    .hasLintViolations(
+                        LintViolation(3, 3, "Missing newline after ';'"),
+                        LintViolation(5, 3, "Missing newline after ';'"),
+                    ).isFormattedAs(formattedCode)
+            }
+
+            @Test
+            fun `Given a multiple init block`() {
+                val code =
+                    """
+                    public class Foo {
+                        init {
+
+                        };init {
+
+                        }
+                    }
+                    """.trimIndent()
+                val formattedCode =
+                    """
+                    public class Foo {
+                        init {
+
+                        }
+                        init {
+
+                        }
+                    }
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .addAdditionalRuleProvider { NoSemicolonsRule() }
+                    .hasLintViolation(4, 7, "Missing newline after ';'")
+                    .isFormattedAs(formattedCode)
+            }
+
+            @Test
+            fun `Given a multiple init block and variables with nested violations`() {
+                val code =
+                    """
+                    public class Foo {
+                        init {
+                            val bar1 = 0; val bar2 = 0;
+                        };init {
+                            val bar3 = 0; val bar4 = 0;
+                        }
+                    }
+                    """.trimIndent()
+                val formattedCode =
+                    """
+                    public class Foo {
+                        init {
+                            val bar1 = 0
+                            val bar2 = 0
+                        }
+                        init {
+                            val bar3 = 0
+                            val bar4 = 0
+                        }
+                    }
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .addAdditionalRuleProvider { NoSemicolonsRule() }
+                    .hasLintViolations(
+                        LintViolation(3, 22, "Missing newline after ';'"),
+                        LintViolation(4, 7, "Missing newline after ';'"),
+                        LintViolation(5, 22, "Missing newline after ';'"),
+                    ).isFormattedAs(formattedCode)
+            }
+        }
+
+        @Nested
+        inner class `Given flow control statements` {
+            @Test
+            fun `Given a multiple for statements`() {
+                val code =
+                    """
+                    fun test() {
+                        for (i in 0..10) {
+                            println(i)
+                        }; for (i in 0..100) {
+                            println(i)
+                        }; for (i in 0..1000) {
+                            println(i)
+                        }
+                    }
+                    """.trimIndent()
+                val formattedCode =
+                    """
+                    fun test() {
+                        for (i in 0..10) {
+                            println(i)
+                        }
+                        for (i in 0..100) {
+                            println(i)
+                        }
+                        for (i in 0..1000) {
+                            println(i)
+                        }
+                    }
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .addAdditionalRuleProvider { NoSemicolonsRule() }
+                    .hasLintViolations(
+                        LintViolation(4, 7, "Missing newline after ';'"),
+                        LintViolation(6, 7, "Missing newline after ';'"),
+                    ).isFormattedAs(formattedCode)
+            }
+
+            @Test
+            fun `Given a multiline while statements`() {
+                val code =
+                    """
+                    fun test() {
+                        while (System.currentTimeMillis() % 2 == 0L) {
+                            println(System.currentTimeMillis())
+                        }; while (Random(System.currentTimeMillis()).nextBoolean()) {
+                            println(System.currentTimeMillis())
+                        }
+                    }
+                    """.trimIndent()
+                val formattedCode =
+                    """
+                    fun test() {
+                        while (System.currentTimeMillis() % 2 == 0L) {
+                            println(System.currentTimeMillis())
+                        }
+                        while (Random(System.currentTimeMillis()).nextBoolean()) {
+                            println(System.currentTimeMillis())
+                        }
+                    }
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .addAdditionalRuleProvider { NoSemicolonsRule() }
+                    .hasLintViolation(4, 7, "Missing newline after ';'")
+                    .isFormattedAs(formattedCode)
+            }
+
+            @Test
+            fun `Given a multiline do-while statements`() {
+                val code =
+                    """
+                    fun test() {
+                        while (System.currentTimeMillis() % 2 == 0L) {
+                            println(System.currentTimeMillis())
+                        }; do {
+                            println(System.currentTimeMillis())
+                        } while (System.currentTimeMillis() % 2 == 0L); do {
+                            println(System.currentTimeMillis())
+                        } while (System.currentTimeMillis() % 2 == 0L)
+                    }
+                    """.trimIndent()
+                val formattedCode =
+                    """
+                    fun test() {
+                        while (System.currentTimeMillis() % 2 == 0L) {
+                            println(System.currentTimeMillis())
+                        }
+                        do {
+                            println(System.currentTimeMillis())
+                        } while (System.currentTimeMillis() % 2 == 0L)
+                        do {
+                            println(System.currentTimeMillis())
+                        } while (System.currentTimeMillis() % 2 == 0L)
+                    }
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .addAdditionalRuleProvider { NoSemicolonsRule() }
+                    .hasLintViolations(
+                        LintViolation(4, 7, "Missing newline after ';'"),
+                        LintViolation(6, 52, "Missing newline after ';'"),
+                    ).isFormattedAs(formattedCode)
+            }
+
+            @Test
+            fun `Given a multiline semi separated control flow with no body`() {
+                val code =
+                    """
+                    fun test() {
+                        for (i in 0..10); for (i in 0..100);while (System.currentTimeMillis() % 2 == 0L); while (Random(System.currentTimeMillis()).nextBoolean());
+                    }
+                    """.trimIndent()
+                val formattedCode =
+                    """
+                    fun test() {
+                        for (i in 0..10);
+                        for (i in 0..100);
+                        while (System.currentTimeMillis() % 2 == 0L);
+                        while (Random(System.currentTimeMillis()).nextBoolean());
+                    }
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .addAdditionalRuleProvider { NoSemicolonsRule() }
+                    .hasLintViolations(
+                        LintViolation(2, 22, "Missing newline after ';'"),
+                        LintViolation(2, 41, "Missing newline after ';'"),
+                        LintViolation(2, 86, "Missing newline after ';'"),
+                    ).isFormattedAs(formattedCode)
+            }
+        }
+
+        @Test
+        fun `Given a multiline semi separated import statement then wrap each expression to a new line`() {
+            val code =
+                """
+                import java.util.ArrayList; import java.util.HashMap
+                """.trimIndent()
+            val formattedCode =
+                """
+                import java.util.ArrayList
+                import java.util.HashMap
+                """.trimIndent()
+            statementWrappingRuleAssertThat(code)
+                .addAdditionalRuleProvider { NoSemicolonsRule() }
+                .hasLintViolation(1, 28, "Missing newline after ';'")
+                .isFormattedAs(formattedCode)
+        }
+
+        @Test
+        fun `Given a multiline semi separated with variables, flow controls and method calls`() {
+            val code =
+                """
+                fun test() {
+                    val a = 0; val b = 0; fun bar() {
+                        // no-op
+                    }; for(i in 0..10) {
+                        println(i); println(i); a++; println(a)
+                    }
+                }
+                """.trimIndent()
+            val formattedCode =
+                """
+                fun test() {
+                    val a = 0
+                    val b = 0
+                    fun bar() {
+                        // no-op
+                    }
+                    for(i in 0..10) {
+                        println(i)
+                        println(i)
+                        a++
+                        println(a)
+                    }
+                }
+                """.trimIndent()
+            statementWrappingRuleAssertThat(code)
+                .addAdditionalRuleProvider { NoSemicolonsRule() }
+                .hasLintViolations(
+                    LintViolation(2, 15, "Missing newline after ';'"),
+                    LintViolation(2, 26, "Missing newline after ';'"),
+                    LintViolation(4, 7, "Missing newline after ';'"),
+                    LintViolation(5, 20, "Missing newline after ';'"),
+                    LintViolation(5, 32, "Missing newline after ';'"),
+                    LintViolation(5, 37, "Missing newline after ';'"),
+                ).isFormattedAs(formattedCode)
+        }
+
+        @Nested
+        inner class `Given enum class` {
+            @Test
+            fun `Given a enum without ending semi`() {
+                val code =
+                    """
+                    enum class FOO1 { ONE, TWO, THREE }
+                    enum class FOO2 {
+                        ONE,
+                        TWO,
+                        THREE
+                    }
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .hasNoLintViolations()
+            }
+
+            @Test
+            fun `Given a enum with ending semi`() {
+                val code =
+                    """
+                    enum class FOO1 { ONE, TWO, THREE; }
+                    enum class FOO2 {
+                        ONE, TWO, THREE;
+                    }
+                    enum class FOO3 {
+                        ONE,
+                        TWO,
+                        THREE;
+                    }
+                    enum class FOO4 {
+                        ONE,
+                        TWO,
+                        THREE,
+                        ;
+                    }
+                    enum class FOO5 {
+                        ONE,
+                        TWO,
+                        THREE,
+                        ;
+                        fun foo() = ""
+                    }
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .hasNoLintViolations()
+            }
+
+            @Test
+            fun `Given a enum with ending semi with comment`() {
+                val code =
+                    """
+                    enum class FOO1 { ONE, TWO, THREE; /* with comment */ }
+                    enum class FOO2 {
+                        ONE,
+                        TWO,
+                        THREE, // single line comment
+                        ; // last single line comment
+                    }
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .hasNoLintViolations()
+            }
+
+            @Test
+            fun `Given enum class with methods`() {
+                val code =
+                    """
+                    enum class FOO {
+                        A, B, C; fun test() = 0
+                    }
+                    """.trimIndent()
+                val formattedCode =
+                    """
+                    enum class FOO {
+                        A, B, C;
+                        fun test() = 0
+                    }
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .addAdditionalRuleProvider { NoSemicolonsRule() }
+                    .hasLintViolation(2, 13, "Missing newline after ';'")
+                    .isFormattedAs(formattedCode)
+            }
+        }
+
+        @Nested
+        inner class `Given companion or object class` {
+            @Test
+            fun `Given a companion object with semicolon and variable`() {
+                val code =
+                    """
+                    class Foo() {
+                        companion object; private var toto: Boolean = false
+                    }
+                    """.trimIndent()
+                val formattedCode =
+                    """
+                    class Foo() {
+                        companion object;
+                        private var toto: Boolean = false
+                    }
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .addAdditionalRuleProvider { NoSemicolonsRule() }
+                    .hasLintViolation(2, 22, "Missing newline after ';'")
+                    .isFormattedAs(formattedCode)
+            }
+
+            @Test
+            fun `Given a companion object with semicolon and comment has not violation`() {
+                val code =
+                    """
+                    class Foo() {
+                        companion object; // single-line comment
+                    }
+                    """.trimIndent()
+                statementWrappingRuleAssertThat(code)
+                    .hasNoLintViolations()
+            }
+        }
+
+        @Test
+        fun `Given a single line block containing multiple statements then reformat block after wrapping the statement`() {
+            val code =
+                """
+                val fooBar =
+                    fooBar()
+                        .map { foo(); bar() }
+                """.trimIndent()
+            val formattedCode =
+                """
+                val fooBar =
+                    fooBar()
+                        .map {
+                            foo()
+                            bar()
+                        }
+                """.trimIndent()
+            statementWrappingRuleAssertThat(code)
+                .addAdditionalRuleProvider { NoSemicolonsRule() }
+                .addAdditionalRuleProvider { IndentationRule() }
+                .hasLintViolations(
+                    LintViolation(3, 16, "Missing newline after '{'"),
+                    LintViolation(3, 22, "Missing newline after ';'"),
+                    LintViolation(3, 29, "Missing newline before '}'"),
+                ).isFormattedAs(formattedCode)
+        }
+
+        @Test
+        fun `Given a single line lambda expression containing multiple statements then reformat block after wrapping the statement`() {
+            val code =
+                """
+                val fooBar =
+                    fooBar()
+                        .map { foo, bar -> print(foo); print(bar) }
+                """.trimIndent()
+            val formattedCode =
+                """
+                val fooBar =
+                    fooBar()
+                        .map { foo, bar ->
+                            print(foo)
+                            print(bar)
+                        }
+                """.trimIndent()
+            statementWrappingRuleAssertThat(code)
+                .addAdditionalRuleProvider { NoSemicolonsRule() }
+                .addAdditionalRuleProvider { IndentationRule() }
+                .hasLintViolations(
+                    LintViolation(3, 28, "Missing newline after '->'"),
+                    LintViolation(3, 39, "Missing newline after ';'"),
+                    LintViolation(3, 51, "Missing newline before '}'"),
+                ).isFormattedAs(formattedCode)
+        }
+
+        @Test
+        fun `Given a single line when entry block containing multiple statements then reformat block after wrapping the statement`() {
+            val code =
+                """
+                val foo =
+                    when (value) {
+                        0 -> { foo(); true }
+                        else -> { bar(); false }
+                    }
+                """.trimIndent()
+            val formattedCode =
+                """
+                val foo =
+                    when (value) {
+                        0 -> {
+                            foo()
+                            true
+                        }
+                        else -> {
+                            bar()
+                            false
+                        }
+                    }
+                """.trimIndent()
+            statementWrappingRuleAssertThat(code)
+                .addAdditionalRuleProvider { NoSemicolonsRule() }
+                .addAdditionalRuleProvider { IndentationRule() }
+                .hasLintViolations(
+                    LintViolation(3, 16, "Missing newline after '{'"),
+                    LintViolation(3, 22, "Missing newline after ';'"),
+                    LintViolation(3, 28, "Missing newline before '}'"),
+                    LintViolation(4, 19, "Missing newline after '{'"),
+                    LintViolation(4, 25, "Missing newline after ';'"),
+                    LintViolation(4, 32, "Missing newline before '}'"),
+                ).isFormattedAs(formattedCode)
+        }
     }
 }
